tools/data-inspector/requirements.txt
pdfplumber==0.11.0
python-docx>=1.1.0
pandas>=2.0.0
openpyxl>=3.1.0
rich>=13.7.0
